reorganization directory
