+++ backend/src/main/java/BackEnd/GestorAlquileres/Auth/DTOs/CustomUserDetails.java
@@ -1,7 +1,7 @@
 package BackEnd.GestorAlquileres.Auth.DTOs;
 
 
-import BackEnd.GestorAlquileres.Users.User;
+import BackEnd.GestorAlquileres.Users.Entities.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
