+++ backend/src/main/java/BackEnd/GestorAlquileres/Auth/Enums/Role.java
@@ -1,0 +1,5 @@
+package BackEnd.GestorAlquileres.Auth.Enums;
+
+public enum Role {
+    USER, ADMIN
+}
