--- backend/src/main/java/BackEnd/GestorAlquileres/Auth/enums/Role.java
@@ -1,5 +1,0 @@
-package BackEnd.GestorAlquileres.Auth.enums;
-
-public enum Role {
-    USER, ADMIN
-}
