--- backend/src/main/java/BackEnd/GestorAlquileres/Auth/services/JwtService.java
@@ -1,37 +1,0 @@
-package BackEnd.GestorAlquileres.Auth.services;
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import BackEnd.GestorAlquileres.Users.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
-
-import java.util.Date;
-
-@Service
-public class JwtService {
-
-    private static final String SECRET_KEY = "clave-super-secreta";
-
-    public String generateToken(User user) {
-        return JWT.create()
-                .withSubject(user.getEmail())
-                .withClaim("id", user.getId())
-                .withClaim("role", user.getRole().name())
-                .withIssuedAt(new Date())
-                .withExpiresAt(new Date(System.currentTimeMillis() + 86400000)) // 1 día
-                .sign(Algorithm.HMAC256(SECRET_KEY));
-    }
-
-    public String extractUsername(String token) {
-        return JWT.require(Algorithm.HMAC256(SECRET_KEY))
-                .build()
-                .verify(token)
-                .getSubject();
-    }
-
-    public boolean isTokenValid(String token, UserDetails userDetails) {
-        String username = extractUsername(token);
-        return username.equals(userDetails.getUsername());
-    }
-}
