--- backend/src/main/java/BackEnd/GestorAlquileres/Cost/services/AuthService.java
@@ -1,4 +1,0 @@
-package BackEnd.GestorAlquileres.Cost.services;
-
-public interface AuthService {
-}
