+++ backend/src/main/java/BackEnd/GestorAlquileres/Exceptions/ErrorResponse.java
@@ -1,0 +1,14 @@
+package BackEnd.GestorAlquileres.Exceptions;
+
+import lombok.*;
+
+import java.util.List;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ErrorResponse {
+    private String errorCode;
+    private String errorMessage;
+    private List<String> details;
+}
