+++ backend/src/main/java/BackEnd/GestorAlquileres/Exceptions/TransactionNotFoundException.java
@@ -1,0 +1,7 @@
+package BackEnd.GestorAlquileres.Exceptions;
+
+public class TransactionNotFoundException extends RuntimeException {
+    public TransactionNotFoundException(String message) {
+        super(message);
+    }
+}
