--- backend/src/main/java/BackEnd/GestorAlquileres/Transactions/services/AuthService.java
@@ -1,4 +1,0 @@
-package BackEnd.GestorAlquileres.Transactions.services;
-
-public interface AuthService {
-}
