--- backend/src/main/java/BackEnd/GestorAlquileres/Users/User.java
@@ -1,31 +1,0 @@
-package BackEnd.GestorAlquileres.Users;
-
-import BackEnd.GestorAlquileres.Auth.enums.Role;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @Column(unique = true, nullable = false)
-    private String username;
-    @Column(unique = true, nullable = false)
-    private String email;
-    @Column(nullable = false)
-    private String password;
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private Role role;
-    private boolean isActive = true;
-
-    public void setIsActive(boolean isActive) {
-        this.isActive = isActive;
-    }
-}
