--- backend/src/main/java/BackEnd/GestorAlquileres/exceptions/ErrorResponse.java
@@ -1,14 +1,0 @@
-package BackEnd.GestorAlquileres.exceptions;
-
-import lombok.*;
-
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class ErrorResponse {
-    private String errorCode;
-    private String errorMessage;
-    private List<String> details;
-}
