--- backend/src/main/java/BackEnd/GestorAlquileres/exceptions/TransactionNotFoundException.java
@@ -1,7 +1,0 @@
-package BackEnd.GestorAlquileres.exceptions;
-
-public class TransactionNotFoundException extends RuntimeException {
-    public TransactionNotFoundException(String message) {
-        super(message);
-    }
-}
